rename: 스켈레톤 컴포넌트 불규칙했던 디렉토리 구조 통일
