+++ frontend/src/components/Skeletons/PinPreviewSkeleton.tsx
