+++ frontend/src/pages/SeeAllNearTopics.tsx
@@ -5,7 +5,7 @@ import useSetNavbarHighlight from '../hooks/useSetNavbarHighlight';
 import Box from '../components/common/Box';
 import Space from '../components/common/Space';
 import Text from '../components/common/Text';
-import TopicCardContainerSkeleton from '../components/TopicCardContainer/TopicCardContainerSkeleton';
+import TopicCardContainerSkeleton from '../components/Skeletons/TopicListSkeleton';
 import { Suspense, lazy } from 'react';
 import useNavigator from '../hooks/useNavigator';
